+++ engines/tensorflow/tensorflow-engine/src/main/java/ai/djl/tensorflow/engine/TfNDArray.java
@@ -1214,7 +1214,7 @@ public class TfNDArray extends NativeResource<TFE_TensorHandle> implements NDArr
 
     /** {@inheritDoc} */
     @Override
-    public NDList unique(long dim, boolean sorted, boolean returnInverse, boolean returnCounts) {
+    public NDList unique(Integer dim, boolean sorted, boolean returnInverse, boolean returnCounts) {
         throw new UnsupportedOperationException("Not implemented yet.");
     }
